feat: merge arc89 JSON into asset metadata
